pyproject.toml
[tool.poetry]
name = "infrastructure"
version = "0.1.0"
description = ""
authors = ["Jack Stockley <34308937+jnstockley@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
httpx = "^0.27.0"
toml = "^0.10.2"
opencv-python = "^4.9.0.80"
testcontainers = "^4.4.0"
pyyaml = "^6.0.1"
docker = "^7.0.0"


[tool.poetry.group.test.dependencies]
pytest = "^8.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
